requirements.txt
disnake
aiohttp
python-dotenv
python-dateutil
pytz
asyncpg